Исправлено  - тестовые модули не содержат ничего кроме тестовых методо
