+++ htmlcov/index.html
@@ -54,7 +54,7 @@
         </h2>
         <p class="text">
             <a class="nav" href="https://coverage.readthedocs.io/en/7.6.4">coverage.py v7.6.4</a>,
-            created at 2024-10-26 22:44 +0300
+            created at 2024-10-28 09:57 +0300
         </p>
     </div>
 </header>
@@ -138,7 +138,7 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io/en/7.6.4">coverage.py v7.6.4</a>,
-            created at 2024-10-26 22:44 +0300
+            created at 2024-10-28 09:57 +0300
         </p>
     </div>
     <aside class="hidden">
