+++ htmlcov/z_c68eb0c7512457e4_burger_py.html
@@ -65,7 +65,7 @@
             <a id="nextFileLink" class="nav" href="z_c68eb0c7512457e4_database_py.html">&#xbb; next</a>
             &nbsp; &nbsp; &nbsp;
             <a class="nav" href="https://coverage.readthedocs.io/en/7.6.4">coverage.py v7.6.4</a>,
-            created at 2024-10-26 18:44 +0300
+            created at 2024-10-28 09:48 +0300
         </p>
         <aside class="hidden">
             <button type="button" class="button_next_chunk" data-shortcut="j"></button>
@@ -137,7 +137,7 @@
             <a class="nav" href="z_c68eb0c7512457e4_database_py.html">&#xbb; next</a>
             &nbsp; &nbsp; &nbsp;
             <a class="nav" href="https://coverage.readthedocs.io/en/7.6.4">coverage.py v7.6.4</a>,
-            created at 2024-10-26 18:44 +0300
+            created at 2024-10-28 09:48 +0300
         </p>
     </div>
 </footer>
