+++ tests/test_burger.py
@@ -1,29 +1,7 @@
 import pytest
 from praktikum.burger import Burger
-from praktikum.bun import Bun
-from praktikum.ingredient import Ingredient
-
-
-class TestInitData:
-    bun = None
-    ingredients = []
-
-
-class TestData:
-    exempl_bun = Bun("exo", 11.22)
-    exempl_ingred = Ingredient('red', 'paper', 9.7)
-    exempl_ingred_sec = Ingredient('ice', 'berry', 9.7)
-
-    @staticmethod
-    def gen_receipt(bun, ingredients_list):
-        receipt = [f'(==== {bun.get_name()} ====)']
-        price = bun.get_price() * 2
-        for ingredient in ingredients_list:
-            receipt.append(f'= {str(ingredient.get_type()).lower()} {ingredient.get_name()} =')
-            price += ingredient.get_price()
-        receipt.append(f'(==== {bun.get_name()} ====)\n')
-        receipt.append(f'Price: {price}')
-        return '\n'.join(receipt)
+from helpers.helpers import BurgerTestData as TestData
+from helpers.helpers import BurgerTestInitData as TestInitData
 
 
 class TestBurger:
