+++ tests/test_database.py
@@ -34,4 +34,3 @@ class TestDatabase:
     def test_available_ingredients(self):
         database = Database()
         assert len(database.available_ingredients()) == 6
-
